[11754] Implement jump (parabolic movement) effect
Also correct destination calculation in Unit::KnockBackFrom - now spline knockback effect works similar to client's effect (same amplitude, speed etc)

+++ src/game/WaypointMovementGenerator.cpp
@@ -312,7 +312,7 @@ void FlightPathMovementGenerator::Reset(Player & player)
 
 bool FlightPathMovementGenerator::Update(Player &player, const uint32 &diff)
 {
-    int32 pointId = player.movespline->currentPathIdx();
+    uint32 pointId = (uint32)player.movespline->currentPathIdx();
     // currentPathIdx returns lastIdx + 1 at arrive
     while (i_currentNode < pointId)
     {
